requirements.txt
tensorflow==2.19.0
flask
numpy
pillow
